requirements.txt
psycopg2-binary>=2.9.0
google-cloud-aiplatform>=1.0.0

